Add ERA5Process and options to download by years

+++ __main__.py
@@ -30,6 +30,9 @@ query_opts.add_option("-f", "--filein", dest="file", action="store",
 query_opts.add_option("-t", "--timeout", dest="timeout", action="store",
     default=60*60*2, help="Waiting time in seconds to stop the process")
 
+# timeout  
+query_opts.add_option("-y", "--byyears", dest="byyears", action="store_true",
+    default=False, help="Start process donwloading by years?")
 
 parser.add_option_group(query_opts)
 (options, args) = parser.parse_args()
@@ -85,12 +88,16 @@ def set_years(ERA5obj,year):
     ERA5obj.year = [year]
     assert isinstance(ERA5obj, Era5Process), "Must be an Era5Process object!"
     
-    
+def options_app(ERA5obj,options):
+    return ERA5Process.byobj(ERA5obj, byyears=options.byyears)
+        
 def split_era5_process(ERA5obj):
     filename = ERA5obj.filename # copy the file name
     years = ERA5obj.year
     times = 0
+    print(f"Process will be split in years {years}\n")
     for iy in years:
+        print(f"Start with year {iy}")
         set_years(ERA5obj, iy)
         times += len(ERA5obj.dates())
         run_era5_process(ERA5obj)
@@ -100,6 +107,7 @@ def split_era5_process(ERA5obj):
             if file_exists(filename):
                 subp.run(['rm',ERA5obj.filename],
                          stdout=subp.PIPE, stderr=subp.PIPE)
+                print(f"File of year {iy} was donwloaded!")
         else:
             subp.run(['cdo','cat',ERA5obj.filename,filename], \
                      stdout=subp.PIPE, stderr=subp.PIPE)
@@ -108,6 +116,11 @@ def split_era5_process(ERA5obj):
             if times == int(cdo_ntime.stdout.decode('utf-8').replace('\n','')):
                 subp.run(['rm',ERA5obj.filename], \
                          stdout=subp.PIPE, stderr=subp.PIPE)
+                print(f"File of year {iy} was donwloaded!")
+                if iy == years[-1]:
+                    print("".center(50,"*"))
+                    print("the downloading process is finished".upper().center(50,"*"))
+                    print("".center(50,"*"))
             else:
                  raise ValueError("ERROR:: time steps do not match!!")
                  
@@ -119,6 +132,7 @@ def main():
         or visit https://code.mpimet.mpg.de/projects/cdo/files
         """
     _Download = downloadERA5_params_from_ini()
+    _Download = options_app(_Download,options)
     print(_Download)
     should_continue = False
     count = 1
@@ -140,26 +154,29 @@ def main():
             continue
         count += 1
     if should_continue:
-        p = mp.Process(target=run_era5_process, name = "run_era5_process", args= (_Download,))
-        p.start()
-        p.join(TIMEOUT)        
-        if p.is_alive():
-            print("Killing the process...\n")
-            p.terminate()
-            p.join()
-            
-            if file_exists(filename):
-                subp.run(['rm',filename], stdout=subp.PIPE, stderr=subp.PIPE)
-            split_era5_process(_Download)      
-        else:
-            if not file_exists(_Download.filename):
-                print(f"ERROR:: {_Download.filename} was not downloaded\n")
-                print("Start process to download year by year and concatenate with CDO\n")
+        if not _Download.byyears:
+            p = mp.Process(target=run_era5_process, name = "run_era5_process", args= (_Download,))
+            p.start()
+            p.join(TIMEOUT)        
+            if p.is_alive():
+                print("Killing the process...\n")
                 p.terminate()
                 p.join()
-                split_era5_process(_Download)
-                
-                   
+                filename = _Download.filename # copy the file name
+                if file_exists(filename):
+                    print(f"{filename} was found, so this will be removed!")
+                    subp.run(['rm',filename], stdout=subp.PIPE, stderr=subp.PIPE)
+                    split_era5_process(_Download)      
+            else:
+                if not file_exists(_Download.filename):
+                    print(f"ERROR:: {_Download.filename} was not downloaded\n")
+                    print("Start process to download year by year and concatenate with CDO\n")
+                    p.terminate()
+                    p.join()
+                    split_era5_process(_Download)
+        else:        
+             print("Download will be done on a yearly basis")
+             split_era5_process(_Download)
 
 if __name__ == "__main__":
